+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add external change listener for SessionStorage and LocalStorage
+
 ## [1.0.2]
 
 ### Fixed